SIH/requirements.txt
streamlit>=1.28.0
Pillow>=9.0.0
requests>=2.28.0
pandas>=1.5.0
plotly>=5.18.0
google-generativeai>=0.3.0
pydeck>=0.8.0



